#48 added more documentation
