+++ .github/workflows/pr-workflow.yml
@@ -33,11 +33,18 @@ jobs:
           key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
           restore-keys: |
             ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+
       - run: yarn install --frozen-lockfile
       - run: yarn run build
+
       - run: yarn hardhat:test
       - name: Publish Unit Test Results
         uses: EnricoMi/publish-unit-test-result-action@v1
         if: always()
         with:
           files: testresult.xml
+
+      - name: Run eslint
+        uses: wearerequired/lint-action@v1
+        with:
+          eslint: true
