+++ src/models/contractTypes.ts
@@ -1,9 +1,12 @@
 import { Contract } from 'ethers';
 
 /**
+ * #### Summary
  * Contracts deployed by hardhat
  * - {chainIds: { networkNames: {contracts} }}, contains an record of contracts
  * - Used by {@link useContractLoader}
+ *
+ * @category Models
  */
 export type TDeployedContracts = {
   [key: string]: {
@@ -16,9 +19,12 @@ export type TDeployedContracts = {
 };
 
 /**
+ * #### Summary
  * A type for external contracts
  * - {chainId: {contracts}}, contains an record of contracts
  * - Used by {@link useContractLoader}
+ *
+ * @category Models
  */
 export type TExternalContracts = {
   [key: number]: {
@@ -29,10 +35,13 @@ export type TExternalContracts = {
 };
 
 /**
+ * #### Summary
  * Contract function information:
  * - contractName
  * - functionname
  * - functionArgs: functionArguments, an array
+ *
+ * @category Models
  */
 export type TContractFunctionInfo = {
   contractName: string;
