+++ src/models/networkTypes.ts
@@ -2,6 +2,7 @@
  * #### Summary
  * A type that describes a network for applications
  *
+ * @category Models
  */
 export type TNetworkInfo = {
   name: string;
